% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_0420.R
\name{IfInRejRegion}
\alias{IfInRejRegion}
\title{FUNCTION: if z statistics fall in the rejection region}
\usage{
IfInRejRegion(rej.region, z)
}
\description{
FUNCTION: if z statistics fall in the rejection region
}
