% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_funs.R
\name{stDesign}
\alias{stDesign}
\title{Get design parameters}
\usage{
stDesign(delta1, delta2, sigma, pi1, method = c("bon", "bon.opt", "mb"),
  alpha = 0.05, beta = 0.2, alpha.min = 1e-04, alpha.input = NULL)
}
\description{
Get design parameters
}
