% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_estimators.R, R/st_tools.R
\name{stEstNaive}
\alias{stEstNaive}
\alias{stEstNaive}
\title{Naive estimator}
\usage{
stEstNaive(data, Y = "Y", A = "A", fml.W = "")

stEstNaive(data, Y = "Y", A = "A", fml.W = "")
}
\description{
Naive estimator

Verify stress test data format
}
