% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_funs.R
\name{stSimu}
\alias{stSimu}
\title{Simulate trials}
\usage{
stSimu(sizen, all.pi, progress = NULL, cnames = c("Design", "Pi", "Bias1",
  "MSE1", "SD1", "Bias2", "MSE2", "SD2", "BiasC", "MSEC", "SDC", "Rej1", "Rej2",
  "RejC", "RejAny"), n.rep = 1000, delta1 = 0, delta2 = 0, n.cores = 4)
}
\description{
Simulate trials
}
